Remove doc as submodule
The sub-directory `doc` contains the documentation
(Markdown files and Doxyfile).
It is no longer a submodule of `PetIBM` (we could not find a consistent
workflow to manage the two repo together).
Once you have done the modifications on the code, update the documentation
in the `doc` directory of the `PetIBM` repo.
Submit the PR, and if it is merge in the `develop` branch,
you can sync the `doc` folder with the GitHub Wiki repo of `PetIBM`.
Thus, the Wiki page only contains the documentation of the `develop` branch.