Modidy data to compare with during regression test
The bug-fix for 0.1.1 concerned the type of the variable holding the value
of the discrete delta function (changed from `int` to `float`).
Therefore, the numerical data we used for comparison during the regression
test for the `TairaColoniusSolver` class is obselete.
Here, we regenerate the data for comparison using the version 0.1.1.