Use CG for GAMG levels
With PETSc-3.7.4, the 2D lid-driven cavity flow runs crashed because of a
indefinite preconditioner for the Poisson solver, which CG cannot handle.
The error only happens when running on multiple procs (and I do not know exactly
why the behavior is different on 1 proc).
A workaround is to change the KSP type on the GAMG levels from the default
Chebyshev to CG using `-poisson_mg_levels_ksp_type cg`.
The discusion with the PETSc developers is available
[here](https://www.mail-archive.com/petsc-users@mcs.anl.gov/msg30221.html).