Add possibility to output files in HDF5 format
We still write the gridlines into the file 'grid.txt' in ASCII format.
However, the method has been moved into the class `CartesianMesh`.
We now call the method in the main function.

The method `CartesianMesh::write()` has been overloaded to write the grid into
a HDF5 file.
The HDF5 grid files (one for each variable) are written when initializing the
fluid solver.

The method `NavierStokesSolver::printInfo()` has been removed.
Now, the methods to print the information about the simulation are called
directly into the main function.

Before this commit, we were only writing the numerical solution in binary
format.
We now add the possibility to output the fields into HDF5 files.
To switch on this mode, the user needs to add the option into the YAML file
`simulationParameters.yaml`: `fileFormat: hdf5`.
(The default behavior is `fileFormat: binary`).

We also add a Python script `createXMFFile.py` to write an XMF file readable
for example by VisIt.
(Use the command-line flag `--help` to display the list of parameters.)