Fix bug in parallel, save grids in separate folder
We write the grids in HDF5 format only when on process with rank 0.
Therefore, we need to use the communicator `PETSC_COMM_SELF` and not
`PETSC_COMM_WORLD` when opening the HDF5 viewer.

We now output the HDF5 grids in a subfolder `grids` only when the file format
is set to `hdf5` in the input file `simulationParameters.yaml`.