Output velocity components into files
In addition to writing the flux components into files, we also write the
velocity components (so that we do not have to calculate the velocity from
fluxes at post-processing stage).

The methods `readVelocities()` and `writeVelocities()` have been added to the
class `NavierStokesSolver`.

The methods `initializeFluxes()` and `initializeLambda()` now only set the
variables `q` and `lambda` with the initial conditions.
The if-branching to decide whether to read the solution from file or not is done
outside the methods `initializeFluxes()` and `initializeLambda()`.

Finally, we now set the variables (`q` and `lambda`) after creating the
operators so that we can use the relationship (q=Ru) to convert the velocity
components into fluxes when reading the velocity form files.