Do not write solution at initial time-step
When restarting a simulation, we read the solution from files, so there is no
need to write the solution immediately after.

When not restarting (i.e., starting a time step 0), writing the initial solution
into files is not necessary (information about the initial conditions are
already present in the input file `flowDescription.yaml`).