Add possibility to write/read flux and/or velocity
We add the possibility to write the flux and/or the velocity variables.
In the input file `simulationParameters.yaml`, the following boolean options
are now available:
- `outputFlux`: write the flux variables into files (default: `true`);
- `outputVelocity`: write the velocity variables into files (default: `false`).

When restarting a simulation from a saved numerical solution, `outputFlux` and
`ouputVelocity` are used to define which files to read from. If both flags are
set to `true`, we choose to read the flux variables from files.