Add abstract class to define type of linear solver
The changes in this commit are based on what @piyueh did in the branch
`AmgXSolvers`.

The new implemented class `Solver` is an abstract class.
The classes `KSPSolver` and `AMGXSolver` inherit from `Solver`.
With these new classes, we move the code related to the linear solver outside
the implementation of the class `NavierStokesSolver`.
With these new classes, we add the possibility to solve a linear system on
multi-GPUs using the AmgX library.

The code related to solving a linear system with AmgX is compiled if you
requested it at configuration time (using the parameters `--with-amgx-dir=PATH`
and `--with-cuda-dir=PATH`).

We now have two new optional parameters for the input file
`simulationParameters.yaml`:
- `vSolveType` with accept the values `CPU` and `GPU`,
- `pSolveType` with accept the values `CPU` and `GPU`.

I still need to test the implementation and to update the documentation.