[WIP] Adding descent restart functionality
* Add `nrestart` parameter (to set up in the input file
`simulationParameters.yaml`) that defines the frequency at which to save the
convective terms into files so that they can be used to restart a simulation.

* Add methods to read and write the convective terms from and to files,
respectively.

* Remove the useless `helpers` methods and add the possibility to display the
matrices and vectors of the problem using the PETSc command-line flags.
I used the function `PetscObjectViewFromOptions` that will output the operator
or vector is a command-line flag is present.

Example: `petibm2d -A_mat_view ascii:A.dat` will save the implicit operator `A`
(after it was set up) into the file `A.dat` in ASCII format.

Other example: `petibm -QTBNQ_mat_view draw -draw_pause -1` will display the
non-zero pattern of the modified Poisson operator into a figure.

I still have to check that the restart is correctly implemented.