Add cli to provide path of input files
The user has now the possibility to provide the path of the YAML
input files in the command-line when running a simulation.

The input files are `cartesianMesh.yaml`, `flowDescription.yaml`,
`simulationParameters.yaml`, and `bodies.yaml`.

Their paths may be set up in the command-line using the using the arguments:
* `-cartesian_mesh <yaml-file-path>`,
* `-flow_description <yaml-file-path>`,
* `-simulation_parameters <yaml-file-path>`,
* `-bodies <yaml-file-path>`.

If those parameters are not provided, PetIBM will look for the regular files
`cartersianMesh.yaml`, `flowDescription.yaml`, `simulationParameter.yaml`,
and `bodies.yaml`, in the simulation directory.