Scatter and update ghosts once per time step
In PetIBM, global vectors are used to obtain the solution of systems using
iterative solvers.
In contrast, local vectors are used to apply stencil operations to calculate
the terms on the right-hand side of the systems.

The IBPM from Taira and Colonius (2007) is a fully-discretized method, in the
sense that the equations are discretized in time and space before the block LU
decomposition is used to obtain the steps of the projection method.
Thus, the boundary conditions should be updated once, at the beginning of each
time step.

This means that scatter from global to local vectors and update of the boundary
ghost values should happen once a time step at the beginning.