Enable mutli-bodies runs (still have to debug)
Note: a simple 2D simulation with two immersed circles showed that the forces
on the first boundary are not correctly calculated.
I still have to find the reason of this bug.

Rename `getBodyInfo` into `setLocalIndexPointsBodies`.
The method calculates the index of all Lagrangian points owned by the local
process.

Move some stuff related to the initialization of the bodies into the method
`initializeBodies` (such as calculating the indices of Eulerian cells that own
a Lagrangian point and calculating the index of Lagrangian points owned by a
local process).

Add method `setLocalNumPhiPoints` that calculates the number of pressure points
owned by the local process.

Adapt method `calculateForcesTC` to enable multi-bodies.
`calculateForcesTC` computes the forces acting on each immersed body by summing
over the corresponding chunk of the Lagrangian force vector (solution of the
modified Poisson system).

Merging the 2D and 3D methods for creating the local to global mapping for the
immersed bodies. (The dimensional specialization was not necessary.)

Adapt method `generateBNQ` to enable multi-bodies. The method creates the
matrices BNQ and ET.

Rename some methods in the class `Body` that describe a little bit better what
they do.
Each body (instance of the class `Body`) is now responsible to keep track of the
Lagrangian points owned by a local process.