Take ET home
This operator ET was only used to spread the Lagrangian forces onto
the Eulerian grid when calculating the forces felt by an immersed
body.
Since we now have a second method to calculate the forces that does
not use ET (we are summing directly over the Lagrangian points), we
removed everything related to the operator ET (as well as the first
method to calculate the forces).

With this commit, we reduce a little bit the memory required to
run a simulation (bye bye ET).