Add possibility to overwrite solvers config files
Using the command-line parameters `-velocity_config_file <file-path>` and/or
`-poisson_config_file <file-path>`, the user has the possibility to provide
different configuration files for the velocity solver and/or the Poisson
solver, respectively.

Note: in the YAML configuration file for the bodies (e.g. `bodies.yaml`),
the path to the body coordinates file is relative to the simulation
directory.