Rewrite method to map local to global indices
Move the methods to register the number of Lagrangian points on each process and
to register the index of Lagrangian points on each process into the `Body`
class.

Add methods to get the total number of Lagrangian points (all boundaries
included) and the number per process in the class `TairaColoniusSolver`.

The local-to-global mapping method (that maps body indices to global indices in
the vector lambda) is now stored in each Body objects.

Write some comments for the attributes of the classes `Body` and
`TairaColoniusSolver`.

Modify the method `generateBNQ` to accounts for the changes listed above.

Remove now-unused methods `setLocalIndexPointsBodies` and
`setLocalNumPhiPoints`.