Fix error for multi-bodies on multi-processes
The method did not work as expected when running a case with multiple
immersed boundaries on multi processes where the Lagrangian forces for
each body are not entirely on the same process.

Problem solved using the number of local points as an offset value and not the
number of points.