Handle periodic boundary conditions for HDF5 grids
When the problem has periodic boundary conditions in one or multiple
directions, the number of points in those directions are augmented by one
for the respective velocity components.

For example, consider a grid nx x ny with periodic boundary conditions
in the y-direction.
The x-velocity field will contain `nx-1 * ny` points but the y-velocity
field will have `nx * ny` points.

We modified the functions to output the different grids in a HDF5 format
to handle this situation.