Pin pressure at one point when using AmgX solver
The Poisson matrix (with Neumman boundary conditions) is singular;
its possesses a non-trivial null space (that consists in the space of
constant functions).

When using Krylov solvers from PETSc, we attach the nullspace to the Poisson
matrix (or its modified version when there is an immersed boundary).
The null space is then removed by the PETSc linear solvers.
This means, when using AmgX as a solver, we do not need to create a
`MatNullSpace` object.
Therefore, the method `setNullSpace` is now called only if using PETSc KSPs.

With AmgX, without removing the null space, the pressure if computed up to a
constant which can vary from one time-step to another.
For example, the farfield pressure for the 2D flow around a cylinder (Re=40)
will change over the course of the simulation.
A workaround is to pin the pressure at one point in the farfield domain.
Currently, the pressure point at the front-bottom-left corner is set to the
value zero.
To do this, we remove all off-diagonal elements in the first row of the
Poisson matrix, while the diagonal element is set to one.
Then, we set the first element of the right-hand side vector to be zero.

Note that, when using AmgX solver, the Poisson matrix is not symmetric anymore.