move to pure velocity based classes and remove lambda properties
Make all classes to a velocity based assumption instead of flux based
or a mixture. It's users' responsibility to make sure whether matrices or the
data in Solution instances are velocity or flux when they use them. To
get flux, users have to manually convert the data in Solution instances
by using R matrix and the member function velocity2Flux, vice versa.

Since the targeting solver in PetIBM is not Taira-Colinius anymore, it
may not useful to always assume a coupling between body force and
pressure in data structure. So now, after separate pressure and body
force, if users' want to have them coupled, they may need to do that
manually.