[bug-fix] Consider all interactions with periodic bcs
Bug caught by @piyueh

Bug only affects simulations involving periodic boundary conditions where
some Lagrangian points are close (less than 0.5 Eulerian grid-spacing)
to a periodic boundary.
In that situation, we have to evaluate the interaction between the Lagrangian
point and the Eulerian nodes on the other side of the periodic boundary.

We remove the first if-statement so that, now, we evaluate all the interactions
between the Lagrangian points and the Eulerian nodes.

We correct the calculation of the length of the domain in each direction.