fix bug: mismatch staggered grid
Due to we don't force the number of local points in each grid (we only
force the distribution of processes to be the same), when we undergo
convection-term calculation, we might miss some values when handling
corner points of a subdomain. For example, a U corner point may require
the value of a ghost corner point from V grid, which STAR stencil doesn't
provide. Therefore, we need to use BOX stencil if we don't want to
write more code to fix it.

In PETSc, BOX stencil doesn't use more memory than STAR stencil. Those
ghost corner points are still allocated by STAR stencil in memory. The
difference between BOX and STAR stencil is that when PETSc does global
to local copying, whether the values of ghost corner points will be
copied.