App: Read and write time value into HDF5 file
Fix #116

The time value is written as an attribute of the pressure dataset into the HDF5
file.
When restarting from a non-zero time-step index (`startStep != 0`), the time
value is read from the HDF5 file that contains the solution.

The methods are implemented in the parent class `NavierStokesSolver` and used
in the applications for IBPM and decoupled IBPM.

The YAML configuration parameters `tstart` was also added to the applications
IBPM and decoupled IBPM and could be used to provide the stating time value.