configure_petsc.m4: test for PETSC_VERSION_RELEASE == 1
Other nonzero values are used to indicate less stable development branches.
Only 1 means a release branch.

I have a more recent version of autotools and thus create thousands of
lines of diffs when I run autogen.sh.  It needs to be run for this
commit to affect builds for users.

Not changed in this commit: PETSc uses semantic versioning so a package
that works with 3.8.0 should also work with any version less than 3.9.
Therefore, I would recommend changing the upper bound to
PETSC_VERSION_LT(3,9,0) if not leaving it open.