Rename TairaColonius to IBPM; update documentation
For consistency with `decoupledibpm`:
* rename class `TairaColoniusSolver` into `IBPMSolver`;
* rename application folder `tairacolonius` into `ibpm`;
* rename executable `petibm-tairacolonius` into `petibm-ibpm`.

Update examples and documentation affected by the renaming.

Update documentation to add more details about the different binary programs
installed by the package.
(Make it explicit that programs work for both 2D and 3D configurations.)

Add references to the documentation for the immersed-boundary methods implemented.

Run `autogen.sh` to update the Makefiles.