Add classes to monitor solution in subregions
* Add abstract class `petibm::misc::ProbeBase` and implementation class
`petibm::misc::ProbeVolume` with code documentation and users documentation.

`ProbeVolume` allows one to monitor the solution of a scalar field variable
(e.g., pressure or one of the velocity components) in a subregion of the
computational domain.

* Add the possibility to configure probes in the PetIBM applications
(Navier-Stokes, IBPM, and decoupled IBPM).
(Configuration done through YAML file.)

* Allow user to build with PETSc 3.8.x to 3.10.x.

* Write a small pre-processing application, `petibm-writemesh`,
that simply creates a mesh and outputs it to a HDF5 file (`grid.h5`).

* Set "solution" as default ouput dir; use logs dir.

The "solution" folder in the simulation directory is the
default directory to save the field solution data.
This directory can be controled using the CLI option
`-output <dir>`.

PETSc logging files are now saved in their own folder.
The "logs" folder in the simulation directory is the
default directory.
This directory can be controled using the CLI option
`-logs <dir>`.