Bugfix: monitor the correct sub-domain with probe
I found a minor bug in the `ProbeVolume` class.
Before this pull-request, the probe was monitoring a field solution
in a sub-domain different than the one configured in the YAML input file.
The bug only affects simulations with more than one MPI process, when we
want to monitor the solution in a sub-volume of the computational domain.
The bug was introduced in version `0.4` of PetIBM.