Fix name of body file to write (for multiple bodies)
This bug in the file name only affected simulations with multiple immersed
bodies with prescribed kinematics.

In cases where multiple immersed bodies are present in the computational
domain, this commit fixes the path of the file to write.
Had to use separate stringstream to keep the time-step index in the filename.