Output all solution files in default solution directory
Before, PetIBM would save the files with the solver iterations, the forces,
and the grid in the "simulation" directory (one that usually contains
the YAML configuration file).
The folder `logs` (that contains PETSc logging files) would also be saved
in the "simulation" directory.
And the field solutions would be saved in the sub-folder `solution` of the
"simulation" directory.
As of now, all these files are saved in the "output" directory.
By default, the output directory is the folder `output` created in the
current working directory.
The user has the possibility to set the path of the output directory
with the command-line option `-output <output_dir>`.
The given path can be either absolute or relative to the current working
directory.

When using probes to monitor the solution at specific points or in a
sub-volume of the domain, the user has to provide the path of the output file
in the YAML configuration file.
This path is now relative to the output directory.

Documentation updated with respect to these changes.